refactor(Total, Gift): 단일책임원칙을 준수하기 위해 Total과 Gift로 나눔
변경 내용:
- Total 클래스가 증정 메뉴를 구하는 기능도 담당하여 이를 분리함
- Gift 클래스를 따로 추가
