+++ src/Total.js
@@ -1,7 +1,5 @@
 import getAllMenus from './utils/menuUtils.js';
 import convertCurrencyUnit from './utils/convertUtils.js';
-import MESSAGE from './constants/message.js';
-import VALUE from './constants/value.js';
 
 const Total = {
   calculateTotal(menu) {
@@ -15,27 +13,8 @@ const Total = {
     return total;
   },
 
-  calculateGift(total) {
-    const type = VALUE.gift.detail.split(' ').at(0);
-    let gift = 0;
-
-    if (total >= VALUE.gift.condition) {
-      gift += VALUE.menu.drink.get(type);
-    }
-
-    return gift;
-  },
-
   displayTotal(total) {
     return convertCurrencyUnit(total);
-  },
-
-  findOutGiftDetail(gift) {
-    if (gift > 0) {
-      return VALUE.gift.detail;
-    }
-
-    return MESSAGE.none;
   },
 };
